Applied hotfix for negative byte bug.
Previously, trying to apply bytes_to_ost to a negative number would
result in an infinite loop, as the program would keep incrementing the
size and ignoring the OverflowError caused by the negative bytes.

The proper solution to this is to rewrite the bytes_to_ostr function so
as to not be so horrible. It's probably possible to figure out what size
the byte array should be without a loop.

Also removed an annoying typo.